extras/previous_versions/v2/azure_ai_foundry/requirements.txt
azure-ai-projects==1.0.0b10
azure-identity
fastapi
uvicorn[standard]